requirements.txt
pandas==2.1.1
numpy==1.24.3
openpyxl==3.1.2
python-dateutil==2.8.2
streamlit==1.27.2
supabase==1.0.3
python-dotenv==1.0.0
pdfplumber==0.10.2
pytesseract==0.3.10
Pillow==10.0.1
plotly==5.17.0 